Move IO functionality to Blocks trait
This allows for more precise control, compared to standard IO traits.
While something similar like write_vectored_at exists, it is platform
specific and unstable. Moreover, the semantics are different from what one
would expect - when stroing data to disk it is syncing it as it goes
along with the writes. This is a more efficient strategy for large writes
and also opens up a possibility for a direct IO implementation

The trait also moves the exact read / full write behavior upstream.

The sync is modified slightly, such that it allows for incomplete writes
to be less than the size of the block, which removes some unnecessary
copying.

These changes have been influenced by the context while writing a
file-backed Blocks implementation.

Removes the need to use advance_iovec, which will likely be reintroduced
later in a file-backed Blocks implementation.