Implement blocks backed by a regular file IO
Includes integration tests, which works within default temp directory of
the OS they running on. The tests cover most of the access patterns to
the file IO. Ideally, there would be a full end-to-end test as well,
which uses the actual streams.

The implementation is generic, but likely won't work on Windows due to
the use of unix-specific API for symlinks when creating locks. There could
be some other less obvious stuff as well.