Start the next file sync after finishing the previous
Although the comments above were stating that this is the case, I have
made made a mistake by calling the asynchronous `sync_file_range` before
waiting for the previous chunk to finish. Fix that.

In addition to that, changed the final blocking sync to wait for both
the previous and the current chunk in one syscall.