Fix a regression when allocating files on slow threads
The problem appears when the thread successfully creates a file with the
index lower than the starting index of the sequence due to a pause after
reading the sequence range. This was detected by accident during a CI build.
The bug can be reproduced by slowing down a single thread with a sleep
right after reading the state in `FileSequence::alloc`.

Since the code relies on syscall atomicity, the solution is a quick hack of
removing the file that has been created out of range.